requirements.txt
Django==4.2.26
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-redis==6.0.0
python-decouple==3.8
requests==2.32.3